Basic Indexing Structure for DataFrames and Series
Index, DataFrame, and Series objects are now genericized by an IIndexLabel-derived type. The IIndexLabel interface is almost a pure marker for now, with only a single specified Equals(IIndexLabel) method as a reminder to override Object.Equals(object).

IDataFrame gains its SeriesCollection fields. SeriesCollection interfaces extend a superinterface that exists solely for indexing consistency, but the subinterfaces are necessary both to abstract a double generic type and to provide typed single-element indexers. The latter function requiring a subinterface is a result of having typed subinterfaces for the different types of Series, which itself is due to intending to support operators for numeric series.

Speaking of, the typed interfaces for 5 types of series now appear. For now, they only have indexers, but as mentioned INumericSeries will gain operator support and other math functions. Other Series types, notably TimeSeries, may gain their own unique functionality.

 Changes to be committed:
	modified:   DataFrame/IDataFrame.cs
	new file:   DataFrame/SeriesCollection/IBoolSeriesCollection.cs
	new file:   DataFrame/SeriesCollection/ICategoricalSeriesCollection.cs
	new file:   DataFrame/SeriesCollection/INumericSeriesCollection.cs
	new file:   DataFrame/SeriesCollection/ISeriesCollection.cs
	new file:   DataFrame/SeriesCollection/IStringSeriesCollection.cs
	new file:   DataFrame/SeriesCollection/ITimeSeriesCollection.cs
	modified:   Index/IIndex.cs
	new file:   Index/IndexLabels/IIndexLabel.cs
	modified:   README.md
	modified:   Series/ISeries.cs
	new file:   Series/TypedSeries/IBoolSeries.cs
	new file:   Series/TypedSeries/ICategoricalSeries.cs
	new file:   Series/TypedSeries/INumericSeries.cs
	new file:   Series/TypedSeries/IStringSeries.cs
	new file:   Series/TypedSeries/ITimeSeries.cs