Parallelization prototype
Should upload the function the series' map functions are referencing. This is an example of separation of concerns - this class should be where all Series parallelization decisions and functionality live.
Haven't decided yet whether utility functions live in a separate top-level project or in Utilities namespaces associated with specific objects. Leaning toward the latter for now, to allow independent customization of thresholds.

 Changes to be committed:
	new file:   Series/Utilities/Parallelization/ParallelizationUtilities.cs