Structural Series & Index Update, with Indexer Iteration
First, for clarity, Index, Series, and Dataframe generic types are now consistently TIndex and TData where present.

ISeries now genericizes on both TIndex and TData (from only TIndex). This allows it to hoist some accessors from the lower-level series interfaces, specifically those which are not benefitted by returning the containing type. For example, accessing a single data element is best done from ISeries, as it doesn't require any functionality specific to a subtype series (and cuts down on code repetition). By contrast, accessing a subset of a, say, numeric series using an enumerable of indices should return a  numeric series object so that it can be further operated upon as a numeric series.
ISeries now also has some indexer-like get/set methods that operate based on row number. These cannot be proper indexers, bexcause they cannot be disambiguated from the TIndex indexer if TIndex is int, which could plausibly happen if one decided to index based on a category.
Indexing logic subject to change when I get to tackling multiindexes.

IndexLabels have been removed. Subsequent testing proved them far too unwieldy to justify using them to fix the above-mentioned disambiguation problem. One expects way row-number-based access should be relatively rare anyway.

IIndex is for now settled on the idea of being a List wrapper with a reverse-lookup dictionary augmentation. Sadly we cannot inherit IList as that pulls in the `T this[int index]` indexer, which can be ambiguous with the `int this[TIndex indexValue]` indexer that is fundamental ot the Index's operation. Wwe also cannot inherit from IDictionary, as that pull in other unwanted functionality. It may be convenient in the future to inherit ICollection, but I have not yet explored it.
I expect to develop a concrete index to directly wrap raw data and a concrete index that wraps a Series (for reindexing DataFrames). I have started in IIndex replicating or ignoring functionality from the pandas API, but realized this was premature, so there is something of a comment manifesto in there for now. Please for now ignore the ValueCounts warning about notnull boxing conversions, the whole area is WIP. A similar manifesto for ISeries got stashed in the Series README, and is probably out of date at time of commit.

Changes to IDataFrame are wholly the result of the generic type rename and the disambiguated accessors.

ISeriesCollection benefitted from similar logic to ISeries. Its former unwieldy generics have been replaced with TIndex and TData, enabled by the new ISeries signature. This allows similar hoisting of some indexers (the ones that do not return typed series). I will note that I am comfortable with a double-generic signature on these two interfaces becuase they are not intended to be used directly.
Some SeriesCollection indexers were discovered to require typed returns and were moved down into the series collection subinterfaces.

ITypedSeries interfaces are updated to match the new indexing structure described for ISeries. Time and Cateorical series remain minimal. Bool and Numeric series have received additional instance members of the Map functions that convert between series types for ease of use. This was discovered in test writing and I think it's cleaner than going to the trouble of moving all the Map methods somewhere they can be defined as extension methods. I also added IStringSeries' versions of the Map functions as well as an elementwise operation for the string Contains method (expect more to follow), again, because there were relevant, logical tests to write.

 Changes to be committed:
	modified:   DataFrame/IDataFrame.cs
	modified:   DataFrame/SeriesCollection/IBoolSeriesCollection.cs
	modified:   DataFrame/SeriesCollection/ICategoricalSeriesCollection.cs
	modified:   DataFrame/SeriesCollection/INumericSeriesCollection.cs
	modified:   DataFrame/SeriesCollection/ISeriesCollection.cs
	modified:   DataFrame/SeriesCollection/IStringSeriesCollection.cs
	modified:   DataFrame/SeriesCollection/ITimeSeriesCollection.cs
	modified:   Index/IIndex.cs
	deleted:    Index/IndexLabels/IIndexLabel.cs
	modified:   Index/README.md
	modified:   Series/ISeries.cs
	modified:   Series/README.md
	modified:   Series/SeriesFactory.cs
	modified:   Series/TypedSeries/IBoolSeries.cs
	modified:   Series/TypedSeries/ICategoricalSeries.cs
	modified:   Series/TypedSeries/INumericSeries.cs
	modified:   Series/TypedSeries/IStringSeries.cs
	modified:   Series/TypedSeries/ITimeSeries.cs